Change: Refactor detail screen file's name
