--- lib/ui/home_features/post_details_screen/local_widgets/detail_section_model.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-
-class DetailSection {
-  DetailSection({
-    this.titleSection,
-    required this.children,
-  });
-
-  final String? titleSection;
-  final List<Widget> children;
-}
