+++ lib/ui/home_features/post_details_screen/local_widgets/post_details_question_input_text_form_field.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+class QuestionInputTextFormField extends StatelessWidget {
+  const QuestionInputTextFormField({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final inputBorder =
+        OutlineInputBorder(borderRadius: BorderRadius.circular(8));
+    return TextFormField(
+      keyboardType: TextInputType.text,
+      decoration: InputDecoration(
+        hintText: 'Type your question here',
+        contentPadding: const EdgeInsets.symmetric(horizontal: 20),
+        border: inputBorder,
+        focusedBorder: inputBorder,
+        enabledBorder: inputBorder,
+      ),
+      // validator: FormValidator().validateEmail,
+      // onSaved: (String value) {
+      //   _loginData.email = value;
+      // },
+    );
+  }
+}
