+++ lib/ui/home_features/post_details_screen/local_widgets/post_details_section_model.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class DetailSection {
+  DetailSection({
+    this.titleSection,
+    required this.children,
+  });
+
+  final String? titleSection;
+  final List<Widget> children;
+}
