+++ lib/ui/home_features/post_details_screen/local_widgets/post_details_separator.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import '../../../../constants/app_colors.dart';
+
+class DetailSeparator extends StatelessWidget {
+  const DetailSeparator({
+    Key? key,
+    required this.height,
+    this.color = AppColors.kScreenBackgroundColor,
+  }) : super(key: key);
+
+  final double height;
+  final Color color;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: height,
+      color: color,
+    );
+  }
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(DoubleProperty('height', height));
+    properties.add(ColorProperty('color', color));
+  }
+}
