+++ lib/ui/home_features/post_details_screen/local_widgets/post_details_title_section.dart
@@ -1,0 +1,96 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:line_icons/line_icons.dart';
+
+import '../../../../models/ui/shared_models/product_model.dart';
+import 'post_details_favourite_toggle_button.dart';
+import 'post_details_section_container.dart';
+import 'post_details_separator.dart';
+
+class TitleSection extends StatelessWidget {
+  const TitleSection({
+    Key? key,
+    required this.product,
+  }) : super(key: key);
+
+  final Product product;
+
+  @override
+  Widget build(BuildContext context) {
+    final size = MediaQuery.of(context).size;
+    return Column(
+      children: [
+        DetailSectionContainer(
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              SizedBox(
+                width: size.width * 0.75,
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      product.title,
+                      style: Theme.of(context).textTheme.headline6,
+                    ),
+                    SizedBox(height: size.height * 0.02),
+                    Text(
+                      '\$${product.price} - \$${product.price + 2}', // demo
+                      style: TextStyle(
+                        fontSize: 18,
+                        color: Theme.of(context).primaryColor,
+                      ),
+                    ),
+                    SizedBox(height: size.height * 0.02),
+                    Text(
+                      product.condition,
+                      style: TextStyle(
+                        fontSize: 18,
+                        fontWeight: FontWeight.w300,
+                        color: ((Theme.of(context).textTheme.bodyText2)!.color)!
+                            .withOpacity(0.8),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              FavouriteToggleButton(isFavourite: product.isFavourite),
+            ],
+          ),
+        ),
+        DetailSeparator(height: size.height * 0.004),
+        DetailSectionContainer(
+          height: 50,
+          child: FittedBox(
+            fit: BoxFit.fitHeight,
+            alignment: Alignment.centerLeft,
+            child: Row(
+              children: [
+                const Icon(
+                  LineIcons.mapMarker,
+                  size: 18,
+                  // color: Theme.of(context).primaryColor,
+                ),
+                SizedBox(width: size.width * 0.01),
+                Text(
+                  product.productLocation,
+                  style: TextStyle(
+                    color: ((Theme.of(context).textTheme.bodyText2)!.color)!
+                        .withOpacity(0.6),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(DiagnosticsProperty<Product>('product', product));
+  }
+}
