+++ lib/ui/home_features/post_details_screen/post_details_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../../../constants/app_colors.dart';
 import '../../../models/arguments/shared/post_details_arguments.dart';
-import 'local_widgets/body.dart';
+import 'local_widgets/post_details_body.dart';
 
 class PostDetailsScreen extends StatelessWidget {
   const PostDetailsScreen({
